Glassfish service name detector (#579)
This is a continuation of #562 that introduces the first of several web
app servers for which we will detect the service name.

Because this is the first (GlassFish), it includes some of the
helper/plumbing utilities. Subsequent additions will mostly be the
`AppServer` implementations (and adding them to the delegate list).