Add buffer handlers and implement buffer metrics (#650)
**Description:**
Add handlers for buffer metrics `process.runtime.jvm.buffer.limit`,
`process.runtime.jvm.buffer.count`, and
`process.runtime.jvm.buffer.usage`.

This was a part of a larger PR here:
https://github.com/open-telemetry/opentelemetry-java-contrib/pull/644.


`jfr-streaming/src/main/java/io/opentelemetry/contrib/jfr/metrics/internal/Constants.java`
has been kept the same as it was in the larger PR because it's used in
all of the smaller constituent PRs. This will hopefully also avoid some
conflicts as the PRs are sequentially merged. As a result, there are
some constants that are defined, but not used in this PR.