Implement GC duration metric (#653)
**Description:**

This PR adds changes that implement the metric
`process.runtime.jvm.gc.duration`.

This was a part of a larger PR here:
https://github.com/open-telemetry/opentelemetry-java-contrib/pull/644.


`jfr-streaming/src/main/java/io/opentelemetry/contrib/jfr/metrics/internal/Constants.java`
has been kept the same as it was in the larger PR because it's used in
all of the smaller constituent PRs. This will hopefully also avoid some
conflicts as the PRs are sequentially merged. As a result, there are
some constants that are defined, but not used in this PR.

**Testing:**

Tests were added but they don't have great coverage. In the future we
need to devise a reliable way to test each GC. Currently, I'm not sure
of a way to change GCs at runtime in order to test them all. I'm also
uncertain of how to reliably force young gen GC and old gen GC. As a
result, the test added in this PR basically just invokes "some" GC and
ensures that the resulting metric has data that is expected.

Co-authored-by: jack-berg <34418638+jack-berg@users.noreply.github.com>