Implement memory metrics (#652)
**Description:**
This PR updates the implementation for the metrics
`process.runtime.jvm.memory.limit` ,
`process.runtime.jvm.memory.usage_after_last_gc` ,
`process.runtime.jvm.memory.committed`,
`process.runtime.jvm.memory.init`, `process.runtime.jvm.memory.usage`


This was a part of a [larger PR
here](https://github.com/open-telemetry/opentelemetry-java-contrib/pull/644).


`jfr-streaming/src/main/java/io/opentelemetry/contrib/jfr/metrics/internal/Constants.java`
has been kept the same as it was in the larger PR because it is used in
all of the smaller constituent PRs. This will hopefully also avoid some
conflicts as the PRs are sequentially merged. As a result, there are
some constants that are defined, but not used in this PR.

Some of the attribute key-value pairs outlined
[here](https://github.com/open-telemetry/opentelemetry-java-instrumentation/tree/main/instrumentation/runtime-metrics/library#garbage-collector-dependent-metrics)
are not present with the current implementation. This is because they
require additional JFR data, or new events. In the future, when JFR is
updated, these handlers should be revisited and updated to cover all the
missing attribute key-value pairs.