JFR features can be enabled and disabled (#709)
Resolves #673.

Groups `RecordedEventHandlers` into "features", which can be toggled on
and off, and which are enabled by default if not already implemented in
the JMX based
[runtime-metrics](https://github.com/open-telemetry/opentelemetry-java-instrumentation/tree/main/instrumentation/runtime-metrics/library).

This was discussed in the 1/5 SIG, as mentioned in this
[comment](https://github.com/open-telemetry/opentelemetry-java-contrib/issues/673#issuecomment-1374223384).

A side effect of this is that the tests can enable the precise set of
features that they test, further increasing isolation.

Also extends `JfrTelemetry#close()` to close all the observable
instruments so that they stop reported data. Does this by having
`RecordedEventHandler` extend `Closeable`, and having each
implementation track its observable instruments.