Fix and enhance resource detection logging. (#711)
I learned (the hard way) today that jul doesn't expand numeric
placeholders in string templates if they are surrounded by single
quotes. Yup.

* `logger.log(INFO, "{0}", "test")` -> `test`
* `logger.log(INFO, "'{0}'", "test")` -> `{0}`

So this results in pretty useless logs here. I also used this as an
opportunity to enhance the details about when/why the resource provider
doesn't kick in.