Breakout ResourceHolder from AwsXrayRemoteSamplerProvider
ResourceHolder provides generic utility for other components in the awsxray package, and could be more broadly used. In this commit, we are breaking out the ResourceHolder inner class to it's own class, and exposing the functionality of getResource, so that other classes can rely on it. Further, we are deprecating the original ResourceHolder, which we believe was not originally intended to be widely visible/accessible.

Related issue: https://github.com/open-telemetry/opentelemetry-java-contrib/issues/789