Breakout ResourceHolder from AwsXrayRemoteSamplerProvider (#801)
**Description:**

ResourceHolder provides generic utility for other components in the
awsxray package, and could be more broadly used. In this commit, we are
breaking out the ResourceHolder inner class to it's own class, and
exposing the functionality of getResource, so that other classes can
rely on it.

Specifically, we are working on new components to meet the needs of
https://github.com/open-telemetry/opentelemetry-java-contrib/issues/789,
which will require the use of this ResourceHolder. Opening up
ResourceHolder seems to us to be the best path forward.

**Existing Issue(s):**

Tangentially related:
https://github.com/open-telemetry/opentelemetry-java-contrib/issues/789

**Testing:**

* Added unit tests, validated they all pass
* `./gradlew clean assemble` succeeds

**Documentation:**

Unclear if any documentation is required here, can help contribute this
as needed. ResourceHolder is a new independent component that can be
used to retrieve the Resource from autoconfiguration.

**Outstanding items:**

None