Update AttributePropagatingSpanProcessorBuilder
In this commit, we are updating AttributePropagatingSpanProcessorBuilder to be somewhat more flexible in use. We still default to the behaviour originally implemented, but we expose the ability to specify span attributes that are used for propagation. This will make the component more reusable, while still achieving the goals required by AWS.
Also, we are renaming Application terminology to Service terminology.