POC for proposed spec change
Because of the following statement added to the spec, the xray and xray-lambda propagators must be combined:

> To avoid potential issues when extracting with an active span context, the xray-lambda propagator SHOULD check if the provided context already has an active span context. If found, the propagator SHOULD return the provided context unmodified.

With this condition, if they are separate the xray-lambda propagator will see the context from xray and avoid modifying the context.
By combining, the propagator is able to consider the context before xray has modified it.